+++ src/domain/registry.ts
@@ -1,5 +1,6 @@
-import { RegistryUrl, unityRegistryUrl } from "./registry-url";
 import { NpmAuth } from "another-npm-registry-client";
+import npmFetch from "npm-registry-fetch";
+import { RegistryUrl, unityRegistryUrl } from "./registry-url";
 
 /**
  * Represents a remote npm-registry.
@@ -20,3 +21,18 @@ export const unityRegistry: Registry = {
   url: unityRegistryUrl,
   auth: null,
 };
+
+/**
+ * Converts a {@link Registry} object into a {@link npmFetch.Options} object for
+ * use in npm registry interactions.
+ * @param registry The registry object to convert.
+ * @returns The created options object.
+ */
+export function makeNpmFetchOptions(registry: Registry): npmFetch.Options {
+  const opts: npmFetch.Options = {
+    registry: registry.url,
+  };
+  const auth = registry.auth;
+  if (auth !== null) Object.assign(opts, auth);
+  return opts;
+}
